rename module files;
