+++ webservice/flaskr/__init__.py
@@ -22,13 +22,16 @@ def create_app(test_config = None):
     from . import db
     db.init_app(app);
 
-    from . import ViolasClientWrapper
-    ViolasClientWrapper.init_app(app);
+    from . import violas_client_wrapper
+    violas_client_wrapper.init_app(app);
 
     from . import wallet
     app.register_blueprint(wallet.bp);
 
     from . import block
-    app.register_blueprint(block.bp)
+    app.register_blueprint(block.bp);
+
+    from . import chain
+    app.register_blueprint(chain.bp);
 
     return app;
